pyproject.toml
[project]
name = "spotify-ensure"
version = "0.1.0"
description = "A tool to ensure invariants of spotify playlists."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "spotipy>=2.25.2",
]
